refactor(website) : include children categories in storefront getAllproducts

+++ Modules/IdentityModule/Identity.Application/Features/AccountManagement/Commands/DeleteRole/DeleteRoleCommandHandler.cs
@@ -19,11 +19,14 @@ namespace Identity.Application.Features.AccountManagement.Commands.DeleteRole
             var role = await _authRepository.GetRoleByIdAsync(request.RoleId, request.TenantId);
             if (role == null)
                 return new DeleteRoleResponse { Success = false, Error = "Role not found in this tenant." };
-            
+     
             // Scope validation
             if (role.Scope != request.Scope)
                 return new DeleteRoleResponse { Success = false, Error = "Unauthorized access to this role scope." };
-
+            if (!string.IsNullOrEmpty(role.Name) &&  role.Name.Contains("SuperAdmin"))
+            {
+                return new DeleteRoleResponse { Success = false, Error = "Prevent removing the admin" };
+            }
             var result = await _authRepository.DeleteRoleAsync(role);
             if (!result.Succeeded)
             {
